add utility class for getting WAD paths
