+++ Core/Resources/IWad/IWadLocator.cs
@@ -1,53 +1,19 @@
-using Microsoft.Win32;
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using Helion.Resources.Archives.Locator;
+using Helion.Resources.Archives;
 
 namespace Helion.Resources.IWad;
 
 public class IWadLocator
 {
-    private static readonly string[] SteamDoomDirs = new[]
-    {
-        "steamapps/common/Ultimate Doom/base",
-        "steamapps/common/Doom 2/base",
-        "steamapps/common/Doom 2/masterbase",
-        "steamapps/common/Doom 2/finaldoombase",
-        "steamapps/common/Final Doom/base",
-        "steamapps/common/DOOM 3 BFG Edition/base/wads",
-    };
-
-    private static readonly string[] LinuxDoomDirs = new[]
-    {
-        "/usr/local/share/doom",
-        "/usr/local/share/games/doom",
-        "/usr/share/doom",
-        "/usr/share/games/doom",
-        "/usr/share/games/doom3bfg/base/wads",
-    };
-
     private readonly List<string> m_directories;
 
     public static IWadLocator CreateDefault(IEnumerable<string> configDirectories)
     {
-        List<string> paths = [Directory.GetCurrentDirectory(), .. configDirectories, .. FilesystemArchiveLocator.GetWadDirsFromEnvVars()];
-
-        var steamPath = GetSteamPath();
-
-        if (Directory.Exists(steamPath))
-        {
-            foreach (var dir in SteamDoomDirs)
-                paths.Add(Path.Combine(steamPath, dir));
-        }
-
-        if (OperatingSystem.IsLinux())
-        {
-            paths.AddRange(LinuxDoomDirs);
-            paths.AddRange(GetLinuxUserPaths());
-        }
-
+        List<string> paths = [Directory.GetCurrentDirectory(), .. configDirectories,
+            .. WadPaths.GetFromSteamAndLinuxDirs(), .. WadPaths.GetFromEnvVars()];
         return new IWadLocator(paths);
     }
 
@@ -89,59 +55,5 @@ public class IWadLocator
             }
         }
         catch { }
-    }
-
-    private static string? GetSteamPath()
-    {
-        if (OperatingSystem.IsWindows())
-        {
-            try
-            {
-                return Registry.GetValue(@"HKEY_CURRENT_USER\Software\Valve\Steam", "SteamPath", null) as string;
-            }
-            catch (Exception)
-            {
-                return Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86), "Steam/");
-            }
-        }
-
-        if (OperatingSystem.IsLinux())
-        {
-            var home = Environment.GetEnvironmentVariable("HOME");
-
-            if (!string.IsNullOrWhiteSpace(home))
-                return $"{home}/.local/share/Steam";
-        }
-
-        return null;
-    }
-
-    private static IEnumerable<string> GetLinuxUserPaths()
-    {
-        var paths = new List<string>();
-
-        var xdgConfigHome = Environment.GetEnvironmentVariable("XDG_CONFIG_HOME");
-
-        if (!string.IsNullOrWhiteSpace(xdgConfigHome))
-            paths.Add($"{xdgConfigHome}/helion");
-
-        var xdgDataHome = Environment.GetEnvironmentVariable("XDG_DATA_HOME");
-
-        if (!string.IsNullOrWhiteSpace(xdgDataHome))
-        {
-            paths.Add($"{xdgDataHome}/doom");
-            paths.Add($"{xdgDataHome}/games/doom");
-        }
-
-        var home = Environment.GetEnvironmentVariable("HOME");
-
-        if (!string.IsNullOrWhiteSpace(home))
-        {
-            paths.Add($"{home}/.config/helion");
-            paths.Add($"{home}/.local/share/doom");
-            paths.Add($"{home}/.local/share/games/doom");
-        }
-
-        return paths;
     }
 }
